// src/types/empresaTypes.ts
export interface Empresa {
  id: number;
  codEmpresa: string;
  cnpj: string;
  razaoSocial: string;
  nomeFantasia: string;
  endereco: string;
  numero: string;
  complemento: string;
  bairro: string;
  cep: string;
  cidade: string;
  uf: string;
  telefone01: string;
  telefone02: string;
  fax: string;
  celular: string;
  whatsapp: string;
  instagram: string;
  linkedin: string;
  nFuncionarios: number;
  dataContribuicao: string;
  valorContribuicao: number;
  dataCadastro: string;
  cadastrante: string;
  observacao: string;
} 